+++ src/App.js
@@ -1,305 +1,15 @@
-import { Toolbar } from "@mui/material";
 import React, { useContext } from "react";
-import MuiAppBar from "@mui/material/AppBar";
-import MenuIcon from "@mui/icons-material/Menu";
-import IconButton from "@mui/material/IconButton";
-import Typography from "@mui/material/Typography";
-import SearchIcon from "@mui/icons-material/Search";
-import { styled, alpha } from "@mui/material/styles";
-import InputBase from "@mui/material/InputBase";
-import Box from "@mui/material/Box";
-import Badge from "@mui/material/Badge";
-import AccountCircle from "@mui/icons-material/AccountCircle";
-import MailIcon from "@mui/icons-material/Mail";
-import NotificationsIcon from "@mui/icons-material/Notifications";
-import MuiDrawer from "@mui/material/Drawer";
-import ListItem from "@mui/material/ListItem";
-import ListItemButton from "@mui/material/ListItemButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListItemText from "@mui/material/ListItemText";
-import List from "@mui/material/List";
-import Home from "@mui/icons-material/Home";
-import History from "@mui/icons-material/History";
-import Subscriptions from "@mui/icons-material/Subscriptions";
-import ThumbUpOffAltIcon from "@mui/icons-material/ThumbUpOffAlt";
-import YouTubeIcon from "@mui/icons-material/YouTube";
-import { Routes, Route, useNavigate } from "react-router-dom";
-import HomePage from "./pages/HomePage";
-import HistoryPage from "./pages/HistoryPage";
-import SubscriptionPage from "./pages/SubscriptionPage";
-import LikedPage from "./pages/LikedPage";
-import VideoPlayer from "./pages/VideoPlayer";
+import MainLayout from "./layout/MainLayout";
 import LoginPage from "./pages/LoginPage";
 import { UserDataContext } from "./context/UserDataContext";
 
-const drawerWidth = 240;
-
-const openedMixin = (theme) => ({
-  width: drawerWidth,
-  transition: theme.transitions.create("width", {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.enteringScreen,
-  }),
-  overflowX: "hidden",
-});
-
-const closedMixin = (theme) => ({
-  transition: theme.transitions.create("width", {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  overflowX: "hidden",
-  width: `calc(${theme.spacing(7)} + 1px)`,
-  [theme.breakpoints.up("sm")]: {
-    width: `calc(${theme.spacing(8)} + 1px)`,
-  },
-});
-
-const DrawerHeader = styled("div")(({ theme }) => ({
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "flex-start",
-  padding: theme.spacing(0, 2),
-  // necessary for content to be below app bar
-  ...theme.mixins.toolbar,
-}));
-
-const Search = styled("div")(({ theme }) => ({
-  position: "relative",
-  borderRadius: 20,
-  backgroundColor: alpha(theme.palette.common.white, 0.09),
-  "&:hover": {
-    backgroundColor: alpha(theme.palette.common.white, 0.1),
-  },
-  marginRight: theme.spacing(2),
-  marginLeft: 0,
-  width: "100%",
-  [theme.breakpoints.up("sm")]: {
-    marginLeft: theme.spacing(30),
-    width: "30%",
-  },
-}));
-
-const SearchIconWrapper = styled("div")(({ theme }) => ({
-  padding: theme.spacing(0, 2),
-  height: "100%",
-  position: "absolute",
-  pointerEvents: "none",
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "center",
-}));
-
-const StyledInputBase = styled(InputBase)(({ theme }) => ({
-  color: "inherit",
-  "& .MuiInputBase-input": {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-    transition: theme.transitions.create("width"),
-    width: "100%",
-    [theme.breakpoints.up("md")]: {
-      width: "20ch",
-    },
-  },
-}));
-
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== "open",
-})(({ theme, open }) => ({
-  backgroundImage: "none",
-  boxShadow: "none",
-  borderBottom: "1px solid rgba(255, 255, 255, 0.12)",
-  zIndex: theme.zIndex.drawer + 1,
-  transition: theme.transitions.create(["width", "margin"], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  // ...(open && {
-  //   marginLeft: drawerWidth,
-  //   width: `calc(100% - ${drawerWidth}px)`,
-  //   transition: theme.transitions.create(["width", "margin"], {
-  //     easing: theme.transitions.easing.sharp,
-  //     duration: theme.transitions.duration.enteringScreen,
-  //   }),
-  // }),
-}));
-
-const Drawer = styled(MuiDrawer, {
-  shouldForwardProp: (prop) => prop !== "open",
-})(({ theme, open }) => ({
-  width: drawerWidth,
-  flexShrink: 0,
-  whiteSpace: "nowrap",
-  boxSizing: "border-box",
-  ...(open && {
-    ...openedMixin(theme),
-    "& .MuiDrawer-paper": openedMixin(theme),
-  }),
-  ...(!open && {
-    ...closedMixin(theme),
-    "& .MuiDrawer-paper": closedMixin(theme),
-  }),
-}));
-
 const App = () => {
-  const user = { userName: "admin ", passWord: "abc@09" };
-  localStorage.setItem("user", JSON.stringify(user));
-  const [open, setOpen] = React.useState(false);
-  const navigate = useNavigate();
+  //conditional rendering for isUserLoggedIn
+
   const [isUserLoggedIn, SetIsUserLoggedIn] = useContext(UserDataContext);
   console.log(isUserLoggedIn);
 
-  const menuList = [
-    {
-      label: "Home",
-      icon: <Home />,
-      path: "/",
-    },
-    {
-      label: "History",
-      icon: <History />,
-      path: "/history",
-    },
-    {
-      label: "Subscriptions",
-      icon: <Subscriptions />,
-      path: "/subscriptions",
-    },
-    {
-      label: "Liked Vidoes",
-      icon: <ThumbUpOffAltIcon />,
-      path: "/liked",
-    },
-  ];
-
-  const handleDrawerOpen = () => {
-    setOpen(!open);
-  };
-
-  const handlePath = (path) => {
-    navigate(path);
-  };
-
-  return (
-    <Box
-      sx={{
-        display: "flex",
-        height: "100vh !important",
-        background: "#ffffff",
-      }}
-    >
-      <AppBar color="primary" position="fixed" open={open}>
-        <Toolbar>
-          <IconButton
-            size="large"
-            edge="start"
-            color="inherit"
-            aria-label="open drawer"
-            onClick={handleDrawerOpen}
-            sx={{
-              marginRight: 5,
-            }}
-          >
-            <MenuIcon />
-          </IconButton>
-          <YouTubeIcon sx={{ color: "red" }} />
-          <Typography
-            variant="h6"
-            noWrap
-            component="div"
-            sx={{ display: { xs: "none", sm: "block" } }}
-          >
-            YouTube
-          </Typography>
-          <Search>
-            <SearchIconWrapper>
-              <SearchIcon />
-            </SearchIconWrapper>
-            <StyledInputBase
-              placeholder="Search…"
-              inputProps={{ "aria-label": "search" }}
-            />
-          </Search>
-          <Box sx={{ flexGrow: 1 }} />
-
-          <Box sx={{ display: { xs: "none", md: "flex" } }}>
-            <IconButton size="large" color="inherit">
-              <Badge color="error">
-                <MailIcon />
-              </Badge>
-            </IconButton>
-            <IconButton size="large" color="inherit">
-              <Badge color="error">
-                <NotificationsIcon />
-              </Badge>
-            </IconButton>
-            <IconButton
-              size="large"
-              edge="end"
-              aria-label="account of current user"
-              aria-haspopup="true"
-              color="inherit"
-            >
-              <AccountCircle />
-            </IconButton>
-          </Box>
-        </Toolbar>
-      </AppBar>
-
-      <Drawer variant="permanent" open={open}>
-        <DrawerHeader />
-        {/* <DrawerHeader>
-          <IconButton onClick={handleDrawerClose}>
-            <MenuIcon />
-          </IconButton>
-        </DrawerHeader> */}
-        <List>
-          {menuList.map((item, index) => (
-            <ListItem key={item.label} disablePadding sx={{ display: "block" }}>
-              <ListItemButton
-                sx={{
-                  // minHeight: 30,
-                  justifyContent: open ? "initial" : "center",
-                  py: 0.8,
-                  width: "90%",
-                  margin: "auto",
-                  borderRadius: 2,
-                }}
-                dense
-                onClick={() => handlePath(item.path)}
-              >
-                <ListItemIcon
-                  sx={{
-                    minWidth: 0,
-                    mr: open ? 3 : "auto",
-                    justifyContent: "center",
-                  }}
-                >
-                  {item.icon}
-                </ListItemIcon>
-                <ListItemText
-                  primary={item.label}
-                  sx={{ opacity: open ? 1 : 0 }}
-                />
-              </ListItemButton>
-            </ListItem>
-          ))}
-        </List>
-      </Drawer>
-      <Box component="main" sx={{ flexGrow: 1, p: 3 }}>
-        <DrawerHeader />
-        <Routes>
-          <Route path="/" element={<HomePage />} />
-          <Route path="/history" element={<HistoryPage />} />
-          <Route path="/subscriptions" element={<SubscriptionPage />} />
-          <Route path="/liked" element={<LikedPage />} />
-          <Route path="/detail" element={<VideoPlayer />} />
-          <Route path="/login" element={<LoginPage />} />
-        </Routes>
-      </Box>
-    </Box>
-  );
+  return <>{isUserLoggedIn ? <MainLayout /> : <LoginPage />}</>;
 };
 
 export default App;
